pyproject.toml
[tool.poetry]
name = "journal-bot"
version = "0.1.0"
description = ""
authors = ["Moses McCall <mosesmc52@gmail.com>"]
readme = "README.md"
packages = [{include = "journal_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
gspread = "^5.12.2"
python-telegram-bot = "^20.7"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
gunicorn = "^21.2.0"
google-auth = "^2.25.2"
google-api-python-client = "^2.111.0"
sqlalchemy = "1.3.19"
sentry-sdk = "^1.39.0"
redis = "^5.0.1"
flower = "^2.0.1"
giphy-client = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
